feat: add disaffiliate staff event request and update user ID types

+++ src/main/java/com/app/teleticket/users/service/UserPhotoStorageService.java
@@ -2,7 +2,7 @@ package com.app.teleticket.users.service;
 
 public interface UserPhotoStorageService {
 
-    String upload(Long userId, String contentType, byte[] bytes);
+    String upload(Integer userId, String contentType, byte[] bytes);
 
     void delete(String key);
 }
